+++ inc/ticket.class.php
@@ -29,6 +29,7 @@
  */
 
 use Glpi\Application\View\TemplateRenderer;
+use Symfony\Component\VarDumper\VarDumper;
 
 if (!defined('GLPI_ROOT')) {
     die("Sorry. You can't access directly to this file");
@@ -108,6 +109,8 @@ class PluginEscaladeTicket
                     }
                 }
             }
+
+            self::removeAssignUsers($item);
             return $item;
         }
     }
@@ -587,7 +590,6 @@ class PluginEscaladeTicket
 
             //delete user
             $ticket_user->delete(['id' => $id]);
-
             if (isset($item->input['_actors'])) {
                 foreach ($item->input['_actors'][$types[$type]] as $key => $actor) {
                     if (
@@ -622,7 +624,6 @@ class PluginEscaladeTicket
         $ticket->getFromDB($tickets_id);
         $groups_id = [];
 
-        self::removeAssignUsers($ticket, $users_id, $type);
 
         // == Add user groups on modification ==
         //check this plugin config